Split platform code for beatmap.cc
